moved NormalizationScheme into normalization file, altered subgraph metadata to be a metadata object not just dict
