+++ Common/kgxmodel.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from Common.biolink_constants import NAMED_THING
-from Common.normalization import NORMALIZATION_CODE_VERSION
+from Common.metadata import GraphMetadata
+from Common.normalization import NormalizationScheme
 
 class kgxnode:
     def __init__(self,
@@ -31,31 +32,6 @@ class kgxedge:
             self.properties = edgeprops
         else:
             self.properties = {}
-
-
-@dataclass
-class NormalizationScheme:
-    node_normalization_version: str = 'latest'
-    edge_normalization_version: str = 'latest'
-    normalization_code_version: str = NORMALIZATION_CODE_VERSION
-    strict: bool = True
-    conflation: bool = False
-
-    def get_composite_normalization_version(self):
-        composite_normalization_version = f'{self.node_normalization_version}_' \
-                                f'{self.edge_normalization_version}_{self.normalization_code_version}'
-        if self.conflation:
-            composite_normalization_version += '_conflated'
-        if self.strict:
-            composite_normalization_version += '_strict'
-        return composite_normalization_version
-
-    def get_metadata_representation(self):
-        return {'node_normalization_version': self.node_normalization_version,
-                'edge_normalization_version': self.edge_normalization_version,
-                'normalization_code_version': self.normalization_code_version,
-                'conflation': self.conflation,
-                'strict': self.strict}
 
 
 @dataclass
@@ -91,13 +67,13 @@ class GraphSource:
 
 @dataclass
 class SubGraphSource(GraphSource):
-    graph_metadata: dict = None
+    graph_metadata: GraphMetadata = None
 
     def get_metadata_representation(self):
         return {'graph_id': self.id,
                 'release_version': self.version,
                 'merge_strategy:': self.merge_strategy,
-                'graph_metadata': self.graph_metadata}
+                'graph_metadata': self.graph_metadata.metadata if self.graph_metadata else None}
 
 
 @dataclass
